backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.11
openai==2.21.0
google-generativeai==0.8.4
python-dotenv==1.2.1
gunicorn==25.1.0
